Add support for cross-compilation
When we are cross compiling, some files are built for the host and some for the
target. We need to differentiate between these scenarios and handle them
differently. This way tests can be built against the native libc, while the
target libc is built separately.

I also removed some unused definitions that were commented out.